// src/lib/generated/Settings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Settings = { schema: number, theme: string, fullscreen: boolean, width: number | null, height: number | null, maxMemoryMb: number, jvmArguments: Array<string>, java25: string | null, java21: string | null, java17: string | null, java8: string | null, downloadConcurrency: number, dataDirectory: string | null, };